requirements.txt
ctransformers==0.2.27
sentence-transformers==2.6.1
pinecone-client==3.1.0
langchain==0.2.3
flask
pypdf
python-dotenv
-e .